ignore: Add support for `.typstignore` files
These files are used as top level ignore files in the `typst-community/
typst-package-template` repository's helper scripts, one should note that the
behavior of the `ignore` crate and the helper scripts would overlap, but not
completely. The `ignore` crate as it is used now ignores hidden directories and
supports reading other ignore files already, which the scripts can't.

This means that, while more correct, the implementation here could be surprising
for users moving from the script solution to typship.